# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Backend Developer Node.js Study Lab API
  version: 1.1.0
  description: >-
    Complete OpenAPI reference for the backend study cases.
    Includes request body schemas, request examples, and response examples
    for both success and common error outcomes.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /cases/15-ci-cd-webhook/github:
    post:
      tags: [Delivery and Architecture]
      summary: Verify GitHub webhook signature using raw request body
      operationId: verifyCase15Webhook
      parameters:
        - $ref: "#/components/parameters/GithubSignatureHeader"
        - $ref: "#/components/parameters/GithubEventHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            examples:
              pushEvent:
                value:
                  ref: refs/heads/main
                  repository:
                    name: study-lab
                  pusher:
                    name: backend-developer
      responses:
        "200":
          description: Signature verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                verified:
                  value:
                    case: 15-ci-cd-webhook
                    verified: true
                    event: push
                    nextStep: Trigger CI/CD deployment workflow
        "400":
          description: Missing or malformed signature header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missingHeader:
                  value:
                    error: Missing x-hub-signature-256 header
                badFormat:
                  value:
                    error: x-hub-signature-256 must use sha256=
        "401":
          description: Signature mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                mismatch:
                  value:
                    case: 15-ci-cd-webhook
                    verified: false
  /cases/15-ci-cd-webhook/signature/generate:
    post:
      tags: [Delivery and Architecture]
      summary: Generate sha256 signature for a payload
      operationId: generateCase15WebhookSignature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            examples:
              samplePayload:
                value:
                  event: push
                  repository: study-lab
      responses:
        "200":
          description: Signature generation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                success:
                  value:
                    case: 15-ci-cd-webhook
                    payload: '{"event":"push","repository":"study-lab"}'
                    signature: sha256=6f0ebf3e0a6fa7a0f83a...
                    headerName: x-hub-signature-256
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    GithubSignatureHeader:
      in: header
      name: x-hub-signature-256
      required: true
      schema:
        type: string
      description: GitHub HMAC digest in the format sha256=<hex>
      example: sha256=6f0ebf3e0a6fa7a0f83a4b91f410f7f5952b0b08356f3b7ec00f2ce39131b615
    GithubEventHeader:
      in: header
      name: x-github-event
      required: false
      schema:
        type: string
      description: GitHub event type
      example: push
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
      required: [error]
...